+++ src/main/java/com/frogedev/hammer_enchant/event/HammerEvents.java
@@ -1,9 +1,7 @@
 package com.frogedev.hammer_enchant.event;
 
 import com.frogedev.hammer_enchant.ModConfig;
-import com.frogedev.hammer_enchant.util.MiningEventHammerHandler;
 import com.frogedev.hammer_enchant.util.MiningHandler;
-import com.frogedev.hammer_enchant.util.UseEventHammerHandler;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
 import net.minecraft.server.level.ServerPlayer;
@@ -48,7 +46,7 @@ public class HammerEvents {
         if(hitResult instanceof BlockHitResult blockHitResult){
             Direction hitDirection = blockHitResult.getDirection();
 
-            MiningEventHammerHandler.MiningEventInfo eventInfo = new MiningEventHammerHandler.MiningEventInfo(event.getPlayer(), event.getPos(), hitDirection);
+            MiningHandler.MiningEventInfo eventInfo = new MiningHandler.MiningEventInfo(event.getPlayer(), event.getPos(), hitDirection);
             if (MiningHandler.INSTANCE.tryPerform(eventInfo)) {
                 event.setCanceled(true);
             }
@@ -73,8 +71,7 @@ public class HammerEvents {
 
         HitResult hitResult =  player.pick(player.getBlockReach(), 0.0f, false);
         if(hitResult instanceof BlockHitResult blockHitResult) {
-            MiningEventHammerHandler.MiningEventInfo eventInfo = new MiningEventHammerHandler.MiningEventInfo(player, breakPos, blockHitResult.getDirection());
-            Iterator<BlockPos> blockPosIter = MiningHandler.INSTANCE.computeCandidatePositionsForFirstQualifyingEvent(Collections.singleton(eventInfo).iterator());
+            Iterator<BlockPos> blockPosIter = MiningHandler.INSTANCE.computeCandidatePositionsForBaseEvent(player, breakPos, blockHitResult.getDirection());
 
             if (!blockPosIter.hasNext()) {
                 return;
